+++ src/main/java/com/gmail/val59000mc/scenarios/scenariolisteners/DoubleOresListener.java
@@ -7,6 +7,7 @@ import com.gmail.val59000mc.utils.OreType;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.block.Block;
+import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.block.BlockBreakEvent;
 import org.bukkit.inventory.ItemStack;
@@ -21,12 +22,14 @@ public class DoubleOresListener extends ScenarioListener {
 			return;
 		}
 
-		if (isEnabled(Scenario.VEIN_MINER)) {
+		final Player player = e.getPlayer();
+		final Block block = e.getBlock();
+
+		if (getScenarioManager().getActiveBlockDropScenario(player, block) != Scenario.DOUBLE_ORES) {
 			return;
 		}
 
-		Block block = e.getBlock();
-		Location loc = e.getBlock().getLocation().add(0.5, 0, 0.5);
+		Location loc = block.getLocation().add(0.5, 0, 0.5);
 		Material type = block.getType();
 		ItemStack drop = null;
 
@@ -36,18 +39,12 @@ public class DoubleOresListener extends ScenarioListener {
 			int xp = oreType.get().getXpPerBlock() * 2;
 			int count = 2;
 
-			if (oreType.get() == OreType.GOLD && isEnabled(Scenario.DOUBLE_GOLD)) {
+			if (OreType.isGold(type) && isEnabled(Scenario.DOUBLE_GOLD)) {
 				count *= 2;
 			}
 
 			drop = oreType.get().getDrop().getStack(count);
 			UhcItems.spawnExtraXp(loc,xp);
-		}
-
-		if (type == Material.SAND) {
-			drop = new ItemStack(Material.GLASS);
-		} else if (type == Material.GRAVEL) {
-			drop = new ItemStack(Material.FLINT);
 		}
 
 		if (drop != null) {